requirements.txt
streamlit==1.28.2
mysql-connector-python==8.2.0
streamlit-js-eval==0.1.5
pillow==10.1.0
pip==23.3.1
pandas==2.0.3
plotly==5.18.0
plotly-express==0.4.1